refactor: split relation of planner into one part.
