+++ datafusion/sql/src/planner.rs
@@ -16,61 +16,25 @@
 // under the License.
 
 //! SQL Query Planner (produces logical plan from SQL AST)
-use std::collections::{HashMap, HashSet};
+use std::collections::HashMap;
 use std::sync::Arc;
 use std::vec;
 
 use arrow_schema::*;
+use sqlparser::ast::ExactNumberInfo;
+use sqlparser::ast::TimezoneInfo;
 use sqlparser::ast::{ColumnDef as SQLColumnDef, ColumnOption};
-use sqlparser::ast::{
-    DataType as SQLDataType, Expr as SQLExpr, Ident, Join, JoinConstraint, JoinOperator,
-    ObjectName, Offset as SQLOffset, Query, Select, SelectItem, SetExpr, SetOperator,
-    ShowCreateObject, ShowStatementFilter, TableAlias, TableFactor, TableWithJoins,
-    UnaryOperator, Value, Values as SQLValues,
-};
-use sqlparser::ast::{ExactNumberInfo, SetQuantifier};
-use sqlparser::ast::{ObjectType, OrderByExpr, Statement};
-use sqlparser::ast::{TimezoneInfo, WildcardAdditionalOptions};
-use sqlparser::parser::ParserError::ParserError;
+use sqlparser::ast::{DataType as SQLDataType, Ident, ObjectName, TableAlias};
 
 use datafusion_common::config::ConfigOptions;
-use datafusion_common::parsers::CompressionTypeVariant;
-use datafusion_common::ToDFSchema;
-use datafusion_common::{
-    field_not_found, Column, DFSchema, DFSchemaRef, DataFusionError, Result, ScalarValue,
-};
+use datafusion_common::{field_not_found, DFSchema, DataFusionError, Result};
 use datafusion_common::{OwnedTableReference, TableReference};
-use datafusion_expr::expr::{Cast, GroupingSet};
-use datafusion_expr::expr_rewriter::normalize_col;
-use datafusion_expr::expr_rewriter::normalize_col_with_schemas;
-use datafusion_expr::logical_plan::builder::project;
-use datafusion_expr::logical_plan::Join as HashJoin;
-use datafusion_expr::logical_plan::JoinConstraint as HashJoinConstraint;
-use datafusion_expr::logical_plan::{
-    Analyze, CreateCatalog, CreateCatalogSchema,
-    CreateExternalTable as PlanCreateExternalTable, CreateMemoryTable, CreateView,
-    DropTable, DropView, Explain, JoinType, LogicalPlan, LogicalPlanBuilder,
-    Partitioning, PlanType, SetVariable, ToStringifiedPlan,
-};
-use datafusion_expr::logical_plan::{Filter, Prepare};
-use datafusion_expr::utils::{
-    expand_qualified_wildcard, expand_wildcard, expr_as_column_expr, expr_to_columns,
-    find_aggregate_exprs, find_column_exprs, find_window_exprs,
-};
-use datafusion_expr::Expr::Alias;
+use datafusion_expr::logical_plan::{LogicalPlan, LogicalPlanBuilder};
+use datafusion_expr::utils::find_column_exprs;
 use datafusion_expr::TableSource;
-use datafusion_expr::{cast, col, lit, AggregateUDF, Expr, ScalarUDF, SubqueryAlias};
+use datafusion_expr::{col, AggregateUDF, Expr, ScalarUDF, SubqueryAlias};
 
-use crate::parser::{CreateExternalTable, DescribeTable, Statement as DFStatement};
-use crate::utils::{make_decimal_type, normalize_ident, resolve_columns};
-
-use super::{
-    parser::DFParser,
-    utils::{
-        check_columns_satisfy_exprs, extract_aliases, rebase_expr,
-        resolve_aliases_to_exprs, resolve_positions_to_exprs,
-    },
-};
+use crate::utils::{make_decimal_type, normalize_ident};
 
 /// The ContextProvider trait allows the query planner to obtain meta-data about tables and
 /// functions referenced in SQL statements
@@ -150,487 +114,7 @@ impl<'a, S: ContextProvider> SqlToRel<'a, S> {
         }
     }
 
-    /// Generate a logical plan from an DataFusion SQL statement
-    pub fn statement_to_plan(&self, statement: DFStatement) -> Result<LogicalPlan> {
-        match statement {
-            DFStatement::CreateExternalTable(s) => self.external_table_to_plan(s),
-            DFStatement::Statement(s) => self.sql_statement_to_plan(*s),
-            DFStatement::DescribeTable(s) => self.describe_table_to_plan(s),
-        }
-    }
-
-    /// Generate a logical plan from an SQL statement
-    pub fn sql_statement_to_plan(&self, statement: Statement) -> Result<LogicalPlan> {
-        self.sql_statement_to_plan_with_context(statement, &mut PlannerContext::new())
-    }
-
-    /// Generate a logical plan from an SQL statement
-    pub fn sql_statement_to_plan_with_context(
-        &self,
-        statement: Statement,
-        planner_context: &mut PlannerContext,
-    ) -> Result<LogicalPlan> {
-        let sql = Some(statement.to_string());
-        match statement {
-            Statement::Explain {
-                verbose,
-                statement,
-                analyze,
-                format: _,
-                describe_alias: _,
-                ..
-            } => self.explain_statement_to_plan(verbose, analyze, *statement),
-            Statement::Query(query) => self.query_to_plan(*query, planner_context),
-            Statement::ShowVariable { variable } => self.show_variable_to_plan(&variable),
-            Statement::SetVariable {
-                local,
-                hivevar,
-                variable,
-                value,
-            } => self.set_variable_to_plan(local, hivevar, &variable, value),
-
-            Statement::CreateTable {
-                query: Some(query),
-                name,
-                columns,
-                constraints,
-                table_properties,
-                with_options,
-                if_not_exists,
-                or_replace,
-                ..
-            } if constraints.is_empty()
-                && table_properties.is_empty()
-                && with_options.is_empty() =>
-            {
-                let plan = self.query_to_plan(*query, planner_context)?;
-                let input_schema = plan.schema();
-
-                let plan = if !columns.is_empty() {
-                    let schema = self.build_schema(columns)?.to_dfschema_ref()?;
-                    if schema.fields().len() != input_schema.fields().len() {
-                        return Err(DataFusionError::Plan(format!(
-                            "Mismatch: {} columns specified, but result has {} columns",
-                            schema.fields().len(),
-                            input_schema.fields().len()
-                        )));
-                    }
-                    let input_fields = input_schema.fields();
-                    let project_exprs = schema
-                        .fields()
-                        .iter()
-                        .zip(input_fields)
-                        .map(|(field, input_field)| {
-                            cast(col(input_field.name()), field.data_type().clone())
-                                .alias(field.name())
-                        })
-                        .collect::<Vec<_>>();
-                    LogicalPlanBuilder::from(plan.clone())
-                        .project(project_exprs)?
-                        .build()?
-                } else {
-                    plan
-                };
-
-                Ok(LogicalPlan::CreateMemoryTable(CreateMemoryTable {
-                    name: object_name_to_table_reference(name)?,
-                    input: Arc::new(plan),
-                    if_not_exists,
-                    or_replace,
-                }))
-            }
-            Statement::CreateView {
-                or_replace,
-                name,
-                columns,
-                query,
-                with_options,
-                ..
-            } if with_options.is_empty() => {
-                let mut plan = self.query_to_plan(*query, &mut PlannerContext::new())?;
-                plan = self.apply_expr_alias(plan, columns)?;
-
-                Ok(LogicalPlan::CreateView(CreateView {
-                    name: object_name_to_table_reference(name)?,
-                    input: Arc::new(plan),
-                    or_replace,
-                    definition: sql,
-                }))
-            }
-            Statement::CreateTable { .. } => Err(DataFusionError::NotImplemented(
-                "Only `CREATE TABLE table_name AS SELECT ...` statement is supported"
-                    .to_string(),
-            )),
-            Statement::ShowCreate { obj_type, obj_name } => match obj_type {
-                ShowCreateObject::Table => self.show_create_table_to_plan(obj_name),
-                _ => Err(DataFusionError::NotImplemented(
-                    "Only `SHOW CREATE TABLE  ...` statement is supported".to_string(),
-                )),
-            },
-            Statement::CreateSchema {
-                schema_name,
-                if_not_exists,
-            } => Ok(LogicalPlan::CreateCatalogSchema(CreateCatalogSchema {
-                schema_name: schema_name.to_string(),
-                if_not_exists,
-                schema: Arc::new(DFSchema::empty()),
-            })),
-            Statement::CreateDatabase {
-                db_name,
-                if_not_exists,
-                ..
-            } => Ok(LogicalPlan::CreateCatalog(CreateCatalog {
-                catalog_name: db_name.to_string(),
-                if_not_exists,
-                schema: Arc::new(DFSchema::empty()),
-            })),
-            Statement::Drop {
-                object_type,
-                if_exists,
-                mut names,
-                cascade: _,
-                restrict: _,
-                purge: _,
-            } => {
-                // We don't support cascade and purge for now.
-                // nor do we support multiple object names
-                let name = match names.len() {
-                    0 => Err(ParserError("Missing table name.".to_string()).into()),
-                    1 => object_name_to_table_reference(names.pop().unwrap()),
-                    _ => {
-                        Err(ParserError("Multiple objects not supported".to_string())
-                            .into())
-                    }
-                }?;
-
-                match object_type {
-                    ObjectType::Table => Ok(LogicalPlan::DropTable(DropTable {
-                        name,
-                        if_exists,
-                        schema: DFSchemaRef::new(DFSchema::empty()),
-                    })),
-                    ObjectType::View => Ok(LogicalPlan::DropView(DropView {
-                        name,
-                        if_exists,
-                        schema: DFSchemaRef::new(DFSchema::empty()),
-                    })),
-                    _ => Err(DataFusionError::NotImplemented(
-                        "Only `DROP TABLE/VIEW  ...` statement is supported currently"
-                            .to_string(),
-                    )),
-                }
-            }
-            Statement::Prepare {
-                name,
-                data_types,
-                statement,
-            } => {
-                // Convert parser data types to DataFusion data types
-                let data_types: Vec<DataType> = data_types
-                    .into_iter()
-                    .map(|t| self.convert_data_type(&t))
-                    .collect::<Result<_>>()?;
-
-                // Create planner context with parameters
-                let mut planner_context =
-                    PlannerContext::new_with_prepare_param_data_types(data_types.clone());
-
-                // Build logical plan for inner statement of the prepare statement
-                let plan = self.sql_statement_to_plan_with_context(
-                    *statement,
-                    &mut planner_context,
-                )?;
-                Ok(LogicalPlan::Prepare(Prepare {
-                    name: name.to_string(),
-                    data_types,
-                    input: Arc::new(plan),
-                }))
-            }
-
-            Statement::ShowTables {
-                extended,
-                full,
-                db_name,
-                filter,
-            } => self.show_tables_to_plan(extended, full, db_name, filter),
-
-            Statement::ShowColumns {
-                extended,
-                full,
-                table_name,
-                filter,
-            } => self.show_columns_to_plan(extended, full, table_name, filter),
-            _ => Err(DataFusionError::NotImplemented(format!(
-                "Unsupported SQL statement: {sql:?}"
-            ))),
-        }
-    }
-
-    /// Generate a logical plan from a "SHOW TABLES" query
-    fn show_tables_to_plan(
-        &self,
-        extended: bool,
-        full: bool,
-        db_name: Option<Ident>,
-        filter: Option<ShowStatementFilter>,
-    ) -> Result<LogicalPlan> {
-        if self.has_table("information_schema", "tables") {
-            // we only support the basic "SHOW TABLES"
-            // https://github.com/apache/arrow-datafusion/issues/3188
-            if db_name.is_some() || filter.is_some() || full || extended {
-                Err(DataFusionError::Plan(
-                    "Unsupported parameters to SHOW TABLES".to_string(),
-                ))
-            } else {
-                let query = "SELECT * FROM information_schema.tables;";
-                let mut rewrite = DFParser::parse_sql(query)?;
-                assert_eq!(rewrite.len(), 1);
-                self.statement_to_plan(rewrite.pop_front().unwrap()) // length of rewrite is 1
-            }
-        } else {
-            Err(DataFusionError::Plan(
-                "SHOW TABLES is not supported unless information_schema is enabled"
-                    .to_string(),
-            ))
-        }
-    }
-
-    /// Generate a logical plan from an SQL query
-    pub fn query_to_plan(
-        &self,
-        query: Query,
-        planner_context: &mut PlannerContext,
-    ) -> Result<LogicalPlan> {
-        self.query_to_plan_with_schema(query, planner_context, None)
-    }
-
-    /// Generate a logical plan from a SQL subquery
-    pub fn subquery_to_plan(
-        &self,
-        query: Query,
-        planner_context: &mut PlannerContext,
-        outer_query_schema: &DFSchema,
-    ) -> Result<LogicalPlan> {
-        self.query_to_plan_with_schema(query, planner_context, Some(outer_query_schema))
-    }
-
-    /// Generate a logic plan from an SQL query.
-    /// It's implementation of `subquery_to_plan` and `query_to_plan`.
-    /// It shouldn't be invoked directly.
-    fn query_to_plan_with_schema(
-        &self,
-        query: Query,
-        planner_context: &mut PlannerContext,
-        outer_query_schema: Option<&DFSchema>,
-    ) -> Result<LogicalPlan> {
-        let set_expr = query.body;
-        if let Some(with) = query.with {
-            // Process CTEs from top to bottom
-            // do not allow self-references
-            if with.recursive {
-                return Err(DataFusionError::NotImplemented(
-                    "Recursive CTEs are not supported".to_string(),
-                ));
-            }
-
-            for cte in with.cte_tables {
-                // A `WITH` block can't use the same name more than once
-                let cte_name = normalize_ident(cte.alias.name.clone());
-                if planner_context.ctes.contains_key(&cte_name) {
-                    return Err(DataFusionError::SQL(ParserError(format!(
-                        "WITH query name {cte_name:?} specified more than once"
-                    ))));
-                }
-                // create logical plan & pass backreferencing CTEs
-                // CTE expr don't need extend outer_query_schema
-                let logical_plan =
-                    self.query_to_plan(*cte.query, &mut planner_context.clone())?;
-
-                // Each `WITH` block can change the column names in the last
-                // projection (e.g. "WITH table(t1, t2) AS SELECT 1, 2").
-                let logical_plan = self.apply_table_alias(logical_plan, cte.alias)?;
-
-                planner_context.ctes.insert(cte_name, logical_plan);
-            }
-        }
-        let plan =
-            self.set_expr_to_plan(*set_expr, planner_context, outer_query_schema)?;
-        let plan = self.order_by(plan, query.order_by)?;
-        self.limit(plan, query.offset, query.limit)
-    }
-
-    fn set_expr_to_plan(
-        &self,
-        set_expr: SetExpr,
-        planner_context: &mut PlannerContext,
-        outer_query_schema: Option<&DFSchema>,
-    ) -> Result<LogicalPlan> {
-        match set_expr {
-            SetExpr::Select(s) => {
-                self.select_to_plan(*s, planner_context, outer_query_schema)
-            }
-            SetExpr::Values(v) => {
-                self.sql_values_to_plan(v, &planner_context.prepare_param_data_types)
-            }
-            SetExpr::SetOperation {
-                op,
-                left,
-                right,
-                set_quantifier,
-            } => {
-                let all = match set_quantifier {
-                    SetQuantifier::All => true,
-                    SetQuantifier::Distinct | SetQuantifier::None => false,
-                };
-
-                let left_plan =
-                    self.set_expr_to_plan(*left, planner_context, outer_query_schema)?;
-                let right_plan =
-                    self.set_expr_to_plan(*right, planner_context, outer_query_schema)?;
-                match (op, all) {
-                    (SetOperator::Union, true) => LogicalPlanBuilder::from(left_plan)
-                        .union(right_plan)?
-                        .build(),
-                    (SetOperator::Union, false) => LogicalPlanBuilder::from(left_plan)
-                        .union_distinct(right_plan)?
-                        .build(),
-                    (SetOperator::Intersect, true) => {
-                        LogicalPlanBuilder::intersect(left_plan, right_plan, true)
-                    }
-                    (SetOperator::Intersect, false) => {
-                        LogicalPlanBuilder::intersect(left_plan, right_plan, false)
-                    }
-                    (SetOperator::Except, true) => {
-                        LogicalPlanBuilder::except(left_plan, right_plan, true)
-                    }
-                    (SetOperator::Except, false) => {
-                        LogicalPlanBuilder::except(left_plan, right_plan, false)
-                    }
-                }
-            }
-            SetExpr::Query(q) => self.query_to_plan(*q, planner_context),
-            _ => Err(DataFusionError::NotImplemented(format!(
-                "Query {set_expr} not implemented yet"
-            ))),
-        }
-    }
-
-    pub fn describe_table_to_plan(
-        &self,
-        statement: DescribeTable,
-    ) -> Result<LogicalPlan> {
-        let DescribeTable { table_name } = statement;
-
-        let where_clause = object_name_to_qualifier(&table_name);
-        let table_ref = object_name_to_table_reference(table_name)?;
-
-        // check if table_name exists
-        let _ = self
-            .schema_provider
-            .get_table_provider((&table_ref).into())?;
-
-        if self.has_table("information_schema", "tables") {
-            let sql = format!(
-                "SELECT column_name, data_type, is_nullable \
-                                FROM information_schema.columns WHERE {where_clause};"
-            );
-            let mut rewrite = DFParser::parse_sql(&sql)?;
-            self.statement_to_plan(rewrite.pop_front().unwrap())
-        } else {
-            Err(DataFusionError::Plan(
-                "DESCRIBE TABLE is not supported unless information_schema is enabled"
-                    .to_string(),
-            ))
-        }
-    }
-
-    /// Generate a logical plan from a CREATE EXTERNAL TABLE statement
-    pub fn external_table_to_plan(
-        &self,
-        statement: CreateExternalTable,
-    ) -> Result<LogicalPlan> {
-        let definition = Some(statement.to_string());
-        let CreateExternalTable {
-            name,
-            columns,
-            file_type,
-            has_header,
-            delimiter,
-            location,
-            table_partition_cols,
-            if_not_exists,
-            file_compression_type,
-            options,
-        } = statement;
-
-        // semantic checks
-        if file_type == "PARQUET" && !columns.is_empty() {
-            Err(DataFusionError::Plan(
-                "Column definitions can not be specified for PARQUET files.".into(),
-            ))?;
-        }
-
-        if file_type != "CSV"
-            && file_type != "JSON"
-            && file_compression_type != CompressionTypeVariant::UNCOMPRESSED
-        {
-            Err(DataFusionError::Plan(
-                "File compression type can be specified for CSV/JSON files.".into(),
-            ))?;
-        }
-
-        let schema = self.build_schema(columns)?;
-
-        // External tables do not support schemas at the moment, so the name is just a table name
-        let name = OwnedTableReference::Bare { table: name };
-
-        Ok(LogicalPlan::CreateExternalTable(PlanCreateExternalTable {
-            schema: schema.to_dfschema_ref()?,
-            name,
-            location,
-            file_type,
-            has_header,
-            delimiter,
-            table_partition_cols,
-            if_not_exists,
-            definition,
-            file_compression_type,
-            options,
-        }))
-    }
-
-    /// Generate a plan for EXPLAIN ... that will print out a plan
-    ///
-    pub fn explain_statement_to_plan(
-        &self,
-        verbose: bool,
-        analyze: bool,
-        statement: Statement,
-    ) -> Result<LogicalPlan> {
-        let plan = self.sql_statement_to_plan(statement)?;
-        let plan = Arc::new(plan);
-        let schema = LogicalPlan::explain_schema();
-        let schema = schema.to_dfschema_ref()?;
-
-        if analyze {
-            Ok(LogicalPlan::Analyze(Analyze {
-                verbose,
-                input: plan,
-                schema,
-            }))
-        } else {
-            let stringified_plans =
-                vec![plan.to_stringified(PlanType::InitialLogicalPlan)];
-            Ok(LogicalPlan::Explain(Explain {
-                verbose,
-                plan,
-                stringified_plans,
-                schema,
-            }))
-        }
-    }
-
-    fn build_schema(&self, columns: Vec<SQLColumnDef>) -> Result<Schema> {
+    pub(crate) fn build_schema(&self, columns: Vec<SQLColumnDef>) -> Result<Schema> {
         let mut fields = Vec::with_capacity(columns.len());
 
         for column in columns {
@@ -649,214 +133,8 @@ impl<'a, S: ContextProvider> SqlToRel<'a, S> {
         Ok(Schema::new(fields))
     }
 
-    fn plan_from_tables(
-        &self,
-        mut from: Vec<TableWithJoins>,
-        planner_context: &mut PlannerContext,
-    ) -> Result<LogicalPlan> {
-        match from.len() {
-            0 => Ok(LogicalPlanBuilder::empty(true).build()?),
-            1 => {
-                let from = from.remove(0);
-                self.plan_table_with_joins(from, planner_context)
-            }
-            _ => {
-                let mut plans = from
-                    .into_iter()
-                    .map(|t| self.plan_table_with_joins(t, planner_context));
-
-                let mut left = LogicalPlanBuilder::from(plans.next().unwrap()?);
-
-                for right in plans {
-                    left = left.cross_join(right?)?;
-                }
-                Ok(left.build()?)
-            }
-        }
-    }
-
-    fn plan_table_with_joins(
-        &self,
-        t: TableWithJoins,
-        planner_context: &mut PlannerContext,
-    ) -> Result<LogicalPlan> {
-        // From clause may exist CTEs, we should separate them from global CTEs.
-        // CTEs in from clause are allowed to be duplicated.
-        // Such as `select * from (WITH source AS (select 1 as e) SELECT * FROM source) t1, (WITH source AS (select 1 as e) SELECT * FROM source) t2;` which is valid.
-        // So always use original global CTEs to plan CTEs in from clause.
-        // Btw, don't need to add CTEs in from to global CTEs.
-        let origin_planner_context = planner_context.clone();
-        let left = self.create_relation(t.relation, planner_context)?;
-        match t.joins.len() {
-            0 => {
-                *planner_context = origin_planner_context;
-                Ok(left)
-            }
-            _ => {
-                let mut joins = t.joins.into_iter();
-                *planner_context = origin_planner_context.clone();
-                let mut left = self.parse_relation_join(
-                    left,
-                    joins.next().unwrap(), // length of joins > 0
-                    planner_context,
-                )?;
-                for join in joins {
-                    *planner_context = origin_planner_context.clone();
-                    left = self.parse_relation_join(left, join, planner_context)?;
-                }
-                *planner_context = origin_planner_context;
-                Ok(left)
-            }
-        }
-    }
-
-    fn parse_relation_join(
-        &self,
-        left: LogicalPlan,
-        join: Join,
-        planner_context: &mut PlannerContext,
-    ) -> Result<LogicalPlan> {
-        let right = self.create_relation(join.relation, planner_context)?;
-        match join.join_operator {
-            JoinOperator::LeftOuter(constraint) => {
-                self.parse_join(left, right, constraint, JoinType::Left, planner_context)
-            }
-            JoinOperator::RightOuter(constraint) => {
-                self.parse_join(left, right, constraint, JoinType::Right, planner_context)
-            }
-            JoinOperator::Inner(constraint) => {
-                self.parse_join(left, right, constraint, JoinType::Inner, planner_context)
-            }
-            JoinOperator::FullOuter(constraint) => {
-                self.parse_join(left, right, constraint, JoinType::Full, planner_context)
-            }
-            JoinOperator::CrossJoin => self.parse_cross_join(left, right),
-            other => Err(DataFusionError::NotImplemented(format!(
-                "Unsupported JOIN operator {other:?}"
-            ))),
-        }
-    }
-
-    fn parse_cross_join(
-        &self,
-        left: LogicalPlan,
-        right: LogicalPlan,
-    ) -> Result<LogicalPlan> {
-        LogicalPlanBuilder::from(left).cross_join(right)?.build()
-    }
-
-    fn parse_join(
-        &self,
-        left: LogicalPlan,
-        right: LogicalPlan,
-        constraint: JoinConstraint,
-        join_type: JoinType,
-        planner_context: &mut PlannerContext,
-    ) -> Result<LogicalPlan> {
-        match constraint {
-            JoinConstraint::On(sql_expr) => {
-                let join_schema = left.schema().join(right.schema())?;
-
-                // parse ON expression
-                let expr = self.sql_to_rex(sql_expr, &join_schema, planner_context)?;
-
-                // ambiguous check
-                ensure_any_column_reference_is_unambiguous(
-                    &expr,
-                    &[left.schema().clone(), right.schema().clone()],
-                )?;
-
-                // normalize all columns in expression
-                let using_columns = expr.to_columns()?;
-                let filter = normalize_col_with_schemas(
-                    expr,
-                    &[left.schema(), right.schema()],
-                    &[using_columns],
-                )?;
-
-                LogicalPlanBuilder::from(left)
-                    .join(
-                        right,
-                        join_type,
-                        (Vec::<Column>::new(), Vec::<Column>::new()),
-                        Some(filter),
-                    )?
-                    .build()
-            }
-            JoinConstraint::Using(idents) => {
-                let keys: Vec<Column> = idents
-                    .into_iter()
-                    .map(|x| Column::from_name(normalize_ident(x)))
-                    .collect();
-                LogicalPlanBuilder::from(left)
-                    .join_using(right, join_type, keys)?
-                    .build()
-            }
-            JoinConstraint::Natural => {
-                // https://issues.apache.org/jira/browse/ARROW-10727
-                Err(DataFusionError::NotImplemented(
-                    "NATURAL JOIN is not supported (https://issues.apache.org/jira/browse/ARROW-10727)".to_string(),
-                ))
-            }
-            JoinConstraint::None => Err(DataFusionError::NotImplemented(
-                "NONE constraint is not supported".to_string(),
-            )),
-        }
-    }
-    fn create_relation(
-        &self,
-        relation: TableFactor,
-        planner_context: &mut PlannerContext,
-    ) -> Result<LogicalPlan> {
-        let (plan, alias) = match relation {
-            TableFactor::Table { name, alias, .. } => {
-                // normalize name and alias
-                let table_ref = object_name_to_table_reference(name)?;
-                let table_name = table_ref.to_string();
-                let cte = planner_context.ctes.get(&table_name);
-                (
-                    match (
-                        cte,
-                        self.schema_provider.get_table_provider((&table_ref).into()),
-                    ) {
-                        (Some(cte_plan), _) => Ok(cte_plan.clone()),
-                        (_, Ok(provider)) => {
-                            LogicalPlanBuilder::scan(&table_name, provider, None)?.build()
-                        }
-                        (None, Err(e)) => Err(e),
-                    }?,
-                    alias,
-                )
-            }
-            TableFactor::Derived {
-                subquery, alias, ..
-            } => {
-                let logical_plan = self.query_to_plan(*subquery, planner_context)?;
-                (logical_plan, alias)
-            }
-            TableFactor::NestedJoin {
-                table_with_joins,
-                alias,
-            } => (
-                self.plan_table_with_joins(*table_with_joins, planner_context)?,
-                alias,
-            ),
-            // @todo Support TableFactory::TableFunction?
-            _ => {
-                return Err(DataFusionError::NotImplemented(format!(
-                    "Unsupported ast node {relation:?} in create_relation"
-                )));
-            }
-        };
-        if let Some(alias) = alias {
-            self.apply_table_alias(plan, alias)
-        } else {
-            Ok(plan)
-        }
-    }
-
     /// Apply the given TableAlias to the top-level projection.
-    fn apply_table_alias(
+    pub(crate) fn apply_table_alias(
         &self,
         plan: LogicalPlan,
         alias: TableAlias,
@@ -869,7 +147,7 @@ impl<'a, S: ContextProvider> SqlToRel<'a, S> {
         self.apply_expr_alias(apply_name_plan, alias.columns)
     }
 
-    fn apply_expr_alias(
+    pub(crate) fn apply_expr_alias(
         &self,
         plan: LogicalPlan,
         idents: Vec<Ident>,
@@ -892,484 +170,8 @@ impl<'a, S: ContextProvider> SqlToRel<'a, S> {
         }
     }
 
-    /// Generate a logic plan from selection clause, the function contain optimization for cross join to inner join
-    /// Related PR: <https://github.com/apache/arrow-datafusion/pull/1566>
-    fn plan_selection(
-        &self,
-        selection: Option<SQLExpr>,
-        plan: LogicalPlan,
-        outer_query_schema: Option<&DFSchema>,
-        planner_context: &mut PlannerContext,
-    ) -> Result<LogicalPlan> {
-        match selection {
-            Some(predicate_expr) => {
-                let mut join_schema = (**plan.schema()).clone();
-                let mut all_schemas: Vec<DFSchemaRef> = vec![];
-                for schema in plan.all_schemas() {
-                    all_schemas.push(schema.clone());
-                }
-                if let Some(outer) = outer_query_schema {
-                    all_schemas.push(Arc::new(outer.clone()));
-                    join_schema.merge(outer);
-                }
-                let x: Vec<&DFSchemaRef> = all_schemas.iter().collect();
-
-                let filter_expr =
-                    self.sql_to_rex(predicate_expr, &join_schema, planner_context)?;
-                let mut using_columns = HashSet::new();
-                expr_to_columns(&filter_expr, &mut using_columns)?;
-                let filter_expr = normalize_col_with_schemas(
-                    filter_expr,
-                    x.as_slice(),
-                    &[using_columns],
-                )?;
-
-                Ok(LogicalPlan::Filter(Filter::try_new(
-                    filter_expr,
-                    Arc::new(plan),
-                )?))
-            }
-            None => Ok(plan),
-        }
-    }
-
-    /// build schema for unqualifier column ambiguous check
-    fn build_schema_for_ambiguous_check(&self, plan: &LogicalPlan) -> Result<DFSchema> {
-        let mut fields = plan.schema().fields().clone();
-
-        let metadata = plan.schema().metadata().clone();
-        if let LogicalPlan::Join(HashJoin {
-            join_constraint: HashJoinConstraint::Using,
-            ref on,
-            ref left,
-            ..
-        }) = plan
-        {
-            // For query: select id from t1 join t2 using(id), this is legal.
-            // We should dedup the fields for cols in using clause.
-            for join_keys in on.iter() {
-                let join_col = &join_keys.0.try_into_col()?;
-                let left_field = left.schema().field_from_column(join_col)?;
-                fields.retain(|field| {
-                    field.unqualified_column().name
-                        != left_field.unqualified_column().name
-                });
-                fields.push(left_field.clone());
-            }
-        }
-
-        DFSchema::new_with_metadata(fields, metadata)
-    }
-
-    /// Generate a logic plan from an SQL select
-    fn select_to_plan(
-        &self,
-        select: Select,
-        planner_context: &mut PlannerContext,
-        outer_query_schema: Option<&DFSchema>,
-    ) -> Result<LogicalPlan> {
-        // check for unsupported syntax first
-        if !select.cluster_by.is_empty() {
-            return Err(DataFusionError::NotImplemented("CLUSTER BY".to_string()));
-        }
-        if !select.lateral_views.is_empty() {
-            return Err(DataFusionError::NotImplemented("LATERAL VIEWS".to_string()));
-        }
-        if select.qualify.is_some() {
-            return Err(DataFusionError::NotImplemented("QUALIFY".to_string()));
-        }
-        if select.top.is_some() {
-            return Err(DataFusionError::NotImplemented("TOP".to_string()));
-        }
-
-        // process `from` clause
-        let plan = self.plan_from_tables(select.from, planner_context)?;
-        let empty_from = matches!(plan, LogicalPlan::EmptyRelation(_));
-        // build from schema for unqualifier column ambiguous check
-        // we should get only one field for unqualifier column from schema.
-        let from_schema = self.build_schema_for_ambiguous_check(&plan)?;
-
-        // process `where` clause
-        let plan = self.plan_selection(
-            select.selection,
-            plan,
-            outer_query_schema,
-            planner_context,
-        )?;
-
-        // process the SELECT expressions, with wildcards expanded.
-        let select_exprs = self.prepare_select_exprs(
-            &plan,
-            select.projection,
-            empty_from,
-            planner_context,
-            &from_schema,
-        )?;
-
-        // having and group by clause may reference aliases defined in select projection
-        let projected_plan = self.project(plan.clone(), select_exprs.clone())?;
-        let mut combined_schema = (**projected_plan.schema()).clone();
-        combined_schema.merge(plan.schema());
-
-        // this alias map is resolved and looked up in both having exprs and group by exprs
-        let alias_map = extract_aliases(&select_exprs);
-
-        // Optionally the HAVING expression.
-        let having_expr_opt = select
-            .having
-            .map::<Result<Expr>, _>(|having_expr| {
-                let having_expr = self.sql_expr_to_logical_expr(
-                    having_expr,
-                    &combined_schema,
-                    planner_context,
-                )?;
-                // This step "dereferences" any aliases in the HAVING clause.
-                //
-                // This is how we support queries with HAVING expressions that
-                // refer to aliased columns.
-                //
-                // For example:
-                //
-                //   SELECT c1, MAX(c2) AS m FROM t GROUP BY c1 HAVING m > 10;
-                //
-                // are rewritten as, respectively:
-                //
-                //   SELECT c1, MAX(c2) AS m FROM t GROUP BY c1 HAVING MAX(c2) > 10;
-                //
-                let having_expr = resolve_aliases_to_exprs(&having_expr, &alias_map)?;
-                normalize_col(having_expr, &projected_plan)
-            })
-            .transpose()?;
-
-        // The outer expressions we will search through for
-        // aggregates. Aggregates may be sourced from the SELECT...
-        let mut aggr_expr_haystack = select_exprs.clone();
-        // ... or from the HAVING.
-        if let Some(having_expr) = &having_expr_opt {
-            aggr_expr_haystack.push(having_expr.clone());
-        }
-
-        // All of the aggregate expressions (deduplicated).
-        let aggr_exprs = find_aggregate_exprs(&aggr_expr_haystack);
-
-        // All of the group by expressions
-        let group_by_exprs = select
-            .group_by
-            .into_iter()
-            .map(|e| {
-                let group_by_expr =
-                    self.sql_expr_to_logical_expr(e, &combined_schema, planner_context)?;
-                // aliases from the projection can conflict with same-named expressions in the input
-                let mut alias_map = alias_map.clone();
-                for f in plan.schema().fields() {
-                    alias_map.remove(f.name());
-                }
-                let group_by_expr = resolve_aliases_to_exprs(&group_by_expr, &alias_map)?;
-                let group_by_expr =
-                    resolve_positions_to_exprs(&group_by_expr, &select_exprs)
-                        .unwrap_or(group_by_expr);
-                let group_by_expr = normalize_col(group_by_expr, &projected_plan)?;
-                self.validate_schema_satisfies_exprs(
-                    plan.schema(),
-                    &[group_by_expr.clone()],
-                )?;
-                Ok(group_by_expr)
-            })
-            .collect::<Result<Vec<Expr>>>()?;
-
-        // process group by, aggregation or having
-        let (plan, mut select_exprs_post_aggr, having_expr_post_aggr) = if !group_by_exprs
-            .is_empty()
-            || !aggr_exprs.is_empty()
-        {
-            self.aggregate(
-                plan,
-                &select_exprs,
-                having_expr_opt.as_ref(),
-                group_by_exprs,
-                aggr_exprs,
-            )?
-        } else {
-            match having_expr_opt {
-                Some(having_expr) => return Err(DataFusionError::Plan(
-                    format!("HAVING clause references: {having_expr} must appear in the GROUP BY clause or be used in an aggregate function"))),
-                None => (plan, select_exprs, having_expr_opt)
-            }
-        };
-
-        let plan = if let Some(having_expr_post_aggr) = having_expr_post_aggr {
-            LogicalPlanBuilder::from(plan)
-                .filter(having_expr_post_aggr)?
-                .build()?
-        } else {
-            plan
-        };
-
-        // process window function
-        let window_func_exprs = find_window_exprs(&select_exprs_post_aggr);
-
-        let plan = if window_func_exprs.is_empty() {
-            plan
-        } else {
-            let plan = LogicalPlanBuilder::window_plan(plan, window_func_exprs.clone())?;
-
-            // re-write the projection
-            select_exprs_post_aggr = select_exprs_post_aggr
-                .iter()
-                .map(|expr| rebase_expr(expr, &window_func_exprs, &plan))
-                .collect::<Result<Vec<Expr>>>()?;
-
-            plan
-        };
-
-        // final projection
-        let plan = project(plan, select_exprs_post_aggr)?;
-
-        // process distinct clause
-        let plan = if select.distinct {
-            LogicalPlanBuilder::from(plan).distinct()?.build()
-        } else {
-            Ok(plan)
-        }?;
-
-        // DISTRIBUTE BY
-        if !select.distribute_by.is_empty() {
-            let x = select
-                .distribute_by
-                .iter()
-                .map(|e| {
-                    self.sql_expr_to_logical_expr(
-                        e.clone(),
-                        &combined_schema,
-                        planner_context,
-                    )
-                })
-                .collect::<Result<Vec<_>>>()?;
-            LogicalPlanBuilder::from(plan)
-                .repartition(Partitioning::DistributeBy(x))?
-                .build()
-        } else {
-            Ok(plan)
-        }
-    }
-
-    /// Returns the `Expr`'s corresponding to a SQL query's SELECT expressions.
-    ///
-    /// Wildcards are expanded into the concrete list of columns.
-    fn prepare_select_exprs(
-        &self,
-        plan: &LogicalPlan,
-        projection: Vec<SelectItem>,
-        empty_from: bool,
-        planner_context: &mut PlannerContext,
-        from_schema: &DFSchema,
-    ) -> Result<Vec<Expr>> {
-        projection
-            .into_iter()
-            .map(|expr| {
-                self.sql_select_to_rex(
-                    expr,
-                    plan,
-                    empty_from,
-                    planner_context,
-                    from_schema,
-                )
-            })
-            .flat_map(|result| match result {
-                Ok(vec) => vec.into_iter().map(Ok).collect(),
-                Err(err) => vec![Err(err)],
-            })
-            .collect::<Result<Vec<Expr>>>()
-    }
-
-    /// Wrap a plan in a projection
-    fn project(&self, input: LogicalPlan, expr: Vec<Expr>) -> Result<LogicalPlan> {
-        self.validate_schema_satisfies_exprs(input.schema(), &expr)?;
-        LogicalPlanBuilder::from(input).project(expr)?.build()
-    }
-
-    /// Create an aggregate plan.
-    ///
-    /// An aggregate plan consists of grouping expressions, aggregate expressions, and an
-    /// optional HAVING expression (which is a filter on the output of the aggregate).
-    ///
-    /// # Arguments
-    ///
-    /// * `input`           - The input plan that will be aggregated. The grouping, aggregate, and
-    ///                       "having" expressions must all be resolvable from this plan.
-    /// * `select_exprs`    - The projection expressions from the SELECT clause.
-    /// * `having_expr_opt` - Optional HAVING clause.
-    /// * `group_by_exprs`  - Grouping expressions from the GROUP BY clause. These can be column
-    ///                       references or more complex expressions.
-    /// * `aggr_exprs`      - Aggregate expressions, such as `SUM(a)` or `COUNT(1)`.
-    ///
-    /// # Return
-    ///
-    /// The return value is a triplet of the following items:
-    ///
-    /// * `plan`                   - A [LogicalPlan::Aggregate] plan for the newly created aggregate.
-    /// * `select_exprs_post_aggr` - The projection expressions rewritten to reference columns from
-    ///                              the aggregate
-    /// * `having_expr_post_aggr`  - The "having" expression rewritten to reference a column from
-    ///                              the aggregate
-    fn aggregate(
-        &self,
-        input: LogicalPlan,
-        select_exprs: &[Expr],
-        having_expr_opt: Option<&Expr>,
-        group_by_exprs: Vec<Expr>,
-        aggr_exprs: Vec<Expr>,
-    ) -> Result<(LogicalPlan, Vec<Expr>, Option<Expr>)> {
-        // create the aggregate plan
-        let plan = LogicalPlanBuilder::from(input.clone())
-            .aggregate(group_by_exprs.clone(), aggr_exprs.clone())?
-            .build()?;
-
-        // in this next section of code we are re-writing the projection to refer to columns
-        // output by the aggregate plan. For example, if the projection contains the expression
-        // `SUM(a)` then we replace that with a reference to a column `SUM(a)` produced by
-        // the aggregate plan.
-
-        // combine the original grouping and aggregate expressions into one list (note that
-        // we do not add the "having" expression since that is not part of the projection)
-        let mut aggr_projection_exprs = vec![];
-        for expr in &group_by_exprs {
-            match expr {
-                Expr::GroupingSet(GroupingSet::Rollup(exprs)) => {
-                    aggr_projection_exprs.extend_from_slice(exprs)
-                }
-                Expr::GroupingSet(GroupingSet::Cube(exprs)) => {
-                    aggr_projection_exprs.extend_from_slice(exprs)
-                }
-                Expr::GroupingSet(GroupingSet::GroupingSets(lists_of_exprs)) => {
-                    for exprs in lists_of_exprs {
-                        aggr_projection_exprs.extend_from_slice(exprs)
-                    }
-                }
-                _ => aggr_projection_exprs.push(expr.clone()),
-            }
-        }
-        aggr_projection_exprs.extend_from_slice(&aggr_exprs);
-
-        // now attempt to resolve columns and replace with fully-qualified columns
-        let aggr_projection_exprs = aggr_projection_exprs
-            .iter()
-            .map(|expr| resolve_columns(expr, &input))
-            .collect::<Result<Vec<Expr>>>()?;
-
-        // next we replace any expressions that are not a column with a column referencing
-        // an output column from the aggregate schema
-        let column_exprs_post_aggr = aggr_projection_exprs
-            .iter()
-            .map(|expr| expr_as_column_expr(expr, &input))
-            .collect::<Result<Vec<Expr>>>()?;
-
-        // next we re-write the projection
-        let select_exprs_post_aggr = select_exprs
-            .iter()
-            .map(|expr| rebase_expr(expr, &aggr_projection_exprs, &input))
-            .collect::<Result<Vec<Expr>>>()?;
-
-        // finally, we have some validation that the re-written projection can be resolved
-        // from the aggregate output columns
-        check_columns_satisfy_exprs(
-            &column_exprs_post_aggr,
-            &select_exprs_post_aggr,
-            "Projection references non-aggregate values",
-        )?;
-
-        // Rewrite the HAVING expression to use the columns produced by the
-        // aggregation.
-        let having_expr_post_aggr = if let Some(having_expr) = having_expr_opt {
-            let having_expr_post_aggr =
-                rebase_expr(having_expr, &aggr_projection_exprs, &input)?;
-
-            check_columns_satisfy_exprs(
-                &column_exprs_post_aggr,
-                &[having_expr_post_aggr.clone()],
-                "HAVING clause references non-aggregate values",
-            )?;
-
-            Some(having_expr_post_aggr)
-        } else {
-            None
-        };
-
-        Ok((plan, select_exprs_post_aggr, having_expr_post_aggr))
-    }
-
-    /// Wrap a plan in a limit
-    fn limit(
-        &self,
-        input: LogicalPlan,
-        skip: Option<SQLOffset>,
-        fetch: Option<SQLExpr>,
-    ) -> Result<LogicalPlan> {
-        if skip.is_none() && fetch.is_none() {
-            return Ok(input);
-        }
-
-        let skip = match skip {
-            Some(skip_expr) => match self.sql_to_rex(
-                skip_expr.value,
-                input.schema(),
-                &mut PlannerContext::new(),
-            )? {
-                Expr::Literal(ScalarValue::Int64(Some(s))) => {
-                    if s < 0 {
-                        return Err(DataFusionError::Plan(format!(
-                            "Offset must be >= 0, '{s}' was provided."
-                        )));
-                    }
-                    Ok(s as usize)
-                }
-                _ => Err(DataFusionError::Plan(
-                    "Unexpected expression in OFFSET clause".to_string(),
-                )),
-            }?,
-            _ => 0,
-        };
-
-        let fetch = match fetch {
-            Some(limit_expr) => {
-                let n = match self.sql_to_rex(
-                    limit_expr,
-                    input.schema(),
-                    &mut PlannerContext::new(),
-                )? {
-                    Expr::Literal(ScalarValue::Int64(Some(n))) => Ok(n as usize),
-                    _ => Err(DataFusionError::Plan(
-                        "Unexpected expression for LIMIT clause".to_string(),
-                    )),
-                }?;
-                Some(n)
-            }
-            _ => None,
-        };
-
-        LogicalPlanBuilder::from(input).limit(skip, fetch)?.build()
-    }
-
-    /// Wrap the logical in a sort
-    fn order_by(
-        &self,
-        plan: LogicalPlan,
-        order_by: Vec<OrderByExpr>,
-    ) -> Result<LogicalPlan> {
-        if order_by.is_empty() {
-            return Ok(plan);
-        }
-
-        let order_by_rex = order_by
-            .into_iter()
-            .map(|e| self.order_by_to_sort_expr(e, plan.schema()))
-            .collect::<Result<Vec<_>>>()?;
-
-        LogicalPlanBuilder::from(plan).sort(order_by_rex)?.build()
-    }
-
     /// Validate the schema provides all of the columns referenced in the expressions.
-    fn validate_schema_satisfies_exprs(
+    pub(crate) fn validate_schema_satisfies_exprs(
         &self,
         schema: &DFSchema,
         exprs: &[Expr],
@@ -1399,379 +201,6 @@ impl<'a, S: ContextProvider> SqlToRel<'a, S> {
                 }),
                 _ => Err(DataFusionError::Internal("Not a column".to_string())),
             })
-    }
-
-    /// ambiguous check for unqualifier column
-    fn column_reference_ambiguous_check(
-        &self,
-        schema: &DFSchema,
-        exprs: &[Expr],
-    ) -> Result<()> {
-        find_column_exprs(exprs)
-            .iter()
-            .try_for_each(|col| match col {
-                Expr::Column(col) => match &col.relation {
-                    None => {
-                        // should get only one field in from_schema.
-                        if schema.fields_with_unqualified_name(&col.name).len() != 1 {
-                            Err(DataFusionError::Internal(format!(
-                                "column reference {} is ambiguous",
-                                col.name
-                            )))
-                        } else {
-                            Ok(())
-                        }
-                    }
-                    _ => Ok(()),
-                },
-                _ => Ok(()),
-            })
-    }
-
-    /// Generate a relational expression from a select SQL expression
-    fn sql_select_to_rex(
-        &self,
-        sql: SelectItem,
-        plan: &LogicalPlan,
-        empty_from: bool,
-        planner_context: &mut PlannerContext,
-        from_schema: &DFSchema,
-    ) -> Result<Vec<Expr>> {
-        match sql {
-            SelectItem::UnnamedExpr(expr) => {
-                let expr = self.sql_to_rex(expr, plan.schema(), planner_context)?;
-                self.column_reference_ambiguous_check(from_schema, &[expr.clone()])?;
-                Ok(vec![normalize_col(expr, plan)?])
-            }
-            SelectItem::ExprWithAlias { expr, alias } => {
-                let select_expr =
-                    self.sql_to_rex(expr, plan.schema(), planner_context)?;
-                self.column_reference_ambiguous_check(
-                    from_schema,
-                    &[select_expr.clone()],
-                )?;
-                let expr = Alias(Box::new(select_expr), normalize_ident(alias));
-                Ok(vec![normalize_col(expr, plan)?])
-            }
-            SelectItem::Wildcard(options) => {
-                Self::check_wildcard_options(options)?;
-
-                if empty_from {
-                    return Err(DataFusionError::Plan(
-                        "SELECT * with no tables specified is not valid".to_string(),
-                    ));
-                }
-                // do not expand from outer schema
-                expand_wildcard(plan.schema().as_ref(), plan)
-            }
-            SelectItem::QualifiedWildcard(ref object_name, options) => {
-                Self::check_wildcard_options(options)?;
-
-                let qualifier = format!("{object_name}");
-                // do not expand from outer schema
-                expand_qualified_wildcard(&qualifier, plan.schema().as_ref(), plan)
-            }
-        }
-    }
-
-    fn check_wildcard_options(options: WildcardAdditionalOptions) -> Result<()> {
-        let WildcardAdditionalOptions {
-            opt_exclude,
-            opt_except,
-            opt_rename,
-        } = options;
-
-        if opt_exclude.is_some() || opt_except.is_some() || opt_rename.is_some() {
-            Err(DataFusionError::NotImplemented(
-                "wildcard * with EXCLUDE, EXCEPT or RENAME not supported ".to_string(),
-            ))
-        } else {
-            Ok(())
-        }
-    }
-
-    /// Generate a relational expression from a SQL expression
-    pub fn sql_to_rex(
-        &self,
-        sql: SQLExpr,
-        schema: &DFSchema,
-        planner_context: &mut PlannerContext,
-    ) -> Result<Expr> {
-        let mut expr = self.sql_expr_to_logical_expr(sql, schema, planner_context)?;
-        expr = self.rewrite_partial_qualifier(expr, schema);
-        self.validate_schema_satisfies_exprs(schema, &[expr.clone()])?;
-        Ok(expr)
-    }
-
-    /// Rewrite aliases which are not-complete (e.g. ones that only include only table qualifier in a schema.table qualified relation)
-    fn rewrite_partial_qualifier(&self, expr: Expr, schema: &DFSchema) -> Expr {
-        match expr {
-            Expr::Column(col) => match &col.relation {
-                Some(q) => {
-                    match schema
-                        .fields()
-                        .iter()
-                        .find(|field| match field.qualifier() {
-                            Some(field_q) => {
-                                field.name() == &col.name
-                                    && field_q.ends_with(&format!(".{q}"))
-                            }
-                            _ => false,
-                        }) {
-                        Some(df_field) => Expr::Column(Column {
-                            relation: df_field.qualifier().cloned(),
-                            name: df_field.name().clone(),
-                        }),
-                        None => Expr::Column(col),
-                    }
-                }
-                None => Expr::Column(col),
-            },
-            _ => expr,
-        }
-    }
-
-    fn sql_values_to_plan(
-        &self,
-        values: SQLValues,
-        param_data_types: &[DataType],
-    ) -> Result<LogicalPlan> {
-        let SQLValues {
-            explicit_row: _,
-            rows,
-        } = values;
-
-        // values should not be based on any other schema
-        let schema = DFSchema::empty();
-        let values = rows
-            .into_iter()
-            .map(|row| {
-                row.into_iter()
-                    .map(|v| match v {
-                        SQLExpr::Value(value) => {
-                            self.parse_value(value, param_data_types)
-                        }
-                        SQLExpr::UnaryOp { op, expr } => self.parse_sql_unary_op(
-                            op,
-                            *expr,
-                            &schema,
-                            &mut PlannerContext::new(),
-                        ),
-                        SQLExpr::BinaryOp { left, op, right } => self
-                            .parse_sql_binary_op(
-                                *left,
-                                op,
-                                *right,
-                                &schema,
-                                &mut PlannerContext::new(),
-                            ),
-                        SQLExpr::TypedString { data_type, value } => {
-                            Ok(Expr::Cast(Cast::new(
-                                Box::new(lit(value)),
-                                self.convert_data_type(&data_type)?,
-                            )))
-                        }
-                        SQLExpr::Cast { expr, data_type } => Ok(Expr::Cast(Cast::new(
-                            Box::new(self.sql_expr_to_logical_expr(
-                                *expr,
-                                &schema,
-                                &mut PlannerContext::new(),
-                            )?),
-                            self.convert_data_type(&data_type)?,
-                        ))),
-                        other => Err(DataFusionError::NotImplemented(format!(
-                            "Unsupported value {other:?} in a values list expression"
-                        ))),
-                    })
-                    .collect::<Result<Vec<_>>>()
-            })
-            .collect::<Result<Vec<_>>>()?;
-        LogicalPlanBuilder::values(values)?.build()
-    }
-
-    fn show_variable_to_plan(&self, variable: &[Ident]) -> Result<LogicalPlan> {
-        let variable = ObjectName(variable.to_vec()).to_string();
-
-        if !self.has_table("information_schema", "df_settings") {
-            return Err(DataFusionError::Plan(
-                "SHOW [VARIABLE] is not supported unless information_schema is enabled"
-                    .to_string(),
-            ));
-        }
-
-        let variable_lower = variable.to_lowercase();
-
-        let query = if variable_lower == "all" {
-            // Add an ORDER BY so the output comes out in a consistent order
-            String::from(
-                "SELECT name, setting FROM information_schema.df_settings ORDER BY name",
-            )
-        } else if variable_lower == "timezone" || variable_lower == "time.zone" {
-            // we could introduce alias in OptionDefinition if this string matching thing grows
-            String::from("SELECT name, setting FROM information_schema.df_settings WHERE name = 'datafusion.execution.time_zone'")
-        } else {
-            format!(
-                "SELECT name, setting FROM information_schema.df_settings WHERE name = '{variable}'"
-            )
-        };
-
-        let mut rewrite = DFParser::parse_sql(&query)?;
-        assert_eq!(rewrite.len(), 1);
-
-        self.statement_to_plan(rewrite.pop_front().unwrap())
-    }
-
-    fn set_variable_to_plan(
-        &self,
-        local: bool,
-        hivevar: bool,
-        variable: &ObjectName,
-        value: Vec<sqlparser::ast::Expr>,
-    ) -> Result<LogicalPlan> {
-        if local {
-            return Err(DataFusionError::NotImplemented(
-                "LOCAL is not supported".to_string(),
-            ));
-        }
-
-        if hivevar {
-            return Err(DataFusionError::NotImplemented(
-                "HIVEVAR is not supported".to_string(),
-            ));
-        }
-
-        let variable = variable.to_string();
-        let mut variable_lower = variable.to_lowercase();
-
-        if variable_lower == "timezone" || variable_lower == "time.zone" {
-            // we could introduce alias in OptionDefinition if this string matching thing grows
-            variable_lower = "datafusion.execution.time_zone".to_string();
-        }
-
-        // parse value string from Expr
-        let value_string = match &value[0] {
-            SQLExpr::Identifier(i) => i.to_string(),
-            SQLExpr::Value(v) => match v {
-                Value::SingleQuotedString(s) => s.to_string(),
-                Value::DollarQuotedString(s) => s.to_string(),
-                Value::Number(_, _) | Value::Boolean(_) => v.to_string(),
-                Value::DoubleQuotedString(_)
-                | Value::UnQuotedString(_)
-                | Value::EscapedStringLiteral(_)
-                | Value::NationalStringLiteral(_)
-                | Value::HexStringLiteral(_)
-                | Value::Null
-                | Value::Placeholder(_) => {
-                    return Err(DataFusionError::Plan(format!(
-                        "Unsupported Value {}",
-                        value[0]
-                    )));
-                }
-            },
-            // for capture signed number e.g. +8, -8
-            SQLExpr::UnaryOp { op, expr } => match op {
-                UnaryOperator::Plus => format!("+{expr}"),
-                UnaryOperator::Minus => format!("-{expr}"),
-                _ => {
-                    return Err(DataFusionError::Plan(format!(
-                        "Unsupported Value {}",
-                        value[0]
-                    )));
-                }
-            },
-            _ => {
-                return Err(DataFusionError::Plan(format!(
-                    "Unsupported Value {}",
-                    value[0]
-                )));
-            }
-        };
-
-        Ok(LogicalPlan::SetVariable(SetVariable {
-            variable: variable_lower,
-            value: value_string,
-            schema: DFSchemaRef::new(DFSchema::empty()),
-        }))
-    }
-
-    fn show_columns_to_plan(
-        &self,
-        extended: bool,
-        full: bool,
-        sql_table_name: ObjectName,
-        filter: Option<ShowStatementFilter>,
-    ) -> Result<LogicalPlan> {
-        if filter.is_some() {
-            return Err(DataFusionError::Plan(
-                "SHOW COLUMNS with WHERE or LIKE is not supported".to_string(),
-            ));
-        }
-
-        if !self.has_table("information_schema", "columns") {
-            return Err(DataFusionError::Plan(
-                "SHOW COLUMNS is not supported unless information_schema is enabled"
-                    .to_string(),
-            ));
-        }
-        // Figure out the where clause
-        let where_clause = object_name_to_qualifier(&sql_table_name);
-
-        // Do a table lookup to verify the table exists
-        let table_ref = object_name_to_table_reference(sql_table_name)?;
-        let _ = self
-            .schema_provider
-            .get_table_provider((&table_ref).into())?;
-
-        // treat both FULL and EXTENDED as the same
-        let select_list = if full || extended {
-            "*"
-        } else {
-            "table_catalog, table_schema, table_name, column_name, data_type, is_nullable"
-        };
-
-        let query = format!(
-            "SELECT {select_list} FROM information_schema.columns WHERE {where_clause}"
-        );
-
-        let mut rewrite = DFParser::parse_sql(&query)?;
-        assert_eq!(rewrite.len(), 1);
-        self.statement_to_plan(rewrite.pop_front().unwrap()) // length of rewrite is 1
-    }
-
-    fn show_create_table_to_plan(
-        &self,
-        sql_table_name: ObjectName,
-    ) -> Result<LogicalPlan> {
-        if !self.has_table("information_schema", "tables") {
-            return Err(DataFusionError::Plan(
-                "SHOW CREATE TABLE is not supported unless information_schema is enabled"
-                    .to_string(),
-            ));
-        }
-        // Figure out the where clause
-        let where_clause = object_name_to_qualifier(&sql_table_name);
-
-        // Do a table lookup to verify the table exists
-        let table_ref = object_name_to_table_reference(sql_table_name)?;
-        let _ = self
-            .schema_provider
-            .get_table_provider((&table_ref).into())?;
-
-        let query = format!(
-            "SELECT table_catalog, table_schema, table_name, definition FROM information_schema.views WHERE {where_clause}"
-        );
-
-        let mut rewrite = DFParser::parse_sql(&query)?;
-        assert_eq!(rewrite.len(), 1);
-        self.statement_to_plan(rewrite.pop_front().unwrap()) // length of rewrite is 1
-    }
-
-    /// Return true if there is a table provider available for "schema.table"
-    fn has_table(&self, schema: &str, table: &str) -> bool {
-        let tables_reference = TableReference::Partial { schema, table };
-        self.schema_provider
-            .get_table_provider(tables_reference)
-            .is_ok()
     }
 
     pub(crate) fn convert_data_type(&self, sql_type: &SQLDataType) -> Result<DataType> {
@@ -1957,65 +386,6 @@ pub fn object_name_to_qualifier(sql_table_name: &ObjectName) -> String {
         })
         .collect::<Vec<_>>()
         .join(" AND ")
-}
-
-/// Ensure any column reference of the expression is unambiguous.
-/// Assume we have two schema:
-/// schema1: a, b ,c
-/// schema2: a, d, e
-///
-/// `schema1.a + schema2.a` is unambiguous.
-/// `a + d` is ambiguous, because `a` may come from schema1 or schema2.
-fn ensure_any_column_reference_is_unambiguous(
-    expr: &Expr,
-    schemas: &[DFSchemaRef],
-) -> Result<()> {
-    if schemas.len() == 1 {
-        return Ok(());
-    }
-    // all referenced columns in the expression that don't have relation
-    let referenced_cols = expr.to_columns()?;
-    let mut no_relation_cols = referenced_cols
-        .iter()
-        .filter_map(|col| {
-            if col.relation.is_none() {
-                Some((col.name.as_str(), 0))
-            } else {
-                None
-            }
-        })
-        .collect::<HashMap<&str, u8>>();
-    // find the name of the column existing in multi schemas.
-    let ambiguous_col_name = schemas
-        .iter()
-        .flat_map(|schema| schema.fields())
-        .map(|field| field.name())
-        .find(|col_name| {
-            no_relation_cols.entry(col_name).and_modify(|v| *v += 1);
-            matches!(
-                no_relation_cols.get_key_value(col_name.as_str()),
-                Some((_, 2..))
-            )
-        });
-
-    if let Some(col_name) = ambiguous_col_name {
-        let maybe_field = schemas
-            .iter()
-            .flat_map(|schema| {
-                schema
-                    .field_with_unqualified_name(col_name)
-                    .map(|f| f.qualified_name())
-                    .ok()
-            })
-            .collect::<Vec<_>>();
-        Err(DataFusionError::Plan(format!(
-            "reference \'{}\' is ambiguous, could be {};",
-            col_name,
-            maybe_field.join(","),
-        )))
-    } else {
-        Ok(())
-    }
 }
 
 #[cfg(test)]
